+++ gdsfactory/__init__.py
@@ -65,7 +65,6 @@ from gdsfactory.add_padding import (
 from gdsfactory.fill import fill_rectangle
 from gdsfactory.pack import pack
 from gdsfactory.grid import grid, grid_with_text
-from gdsfactory.generic_tech import LAYER, LAYER_STACK, get_generic_pdk
 from gdsfactory.pdk import (
     Pdk,
     get_component,
@@ -87,8 +86,6 @@ __all__ = (
     "ComponentReference",
     "CrossSection",
     "Group",
-    "LAYER",
-    "LAYER_STACK",
     "Label",
     "Path",
     "Pdk",
@@ -126,7 +123,6 @@ __all__ = (
     "get_cross_section_factories",
     "get_layer",
     "get_padding_points",
-    "get_generic_pdk",
     "grid",
     "grid_with_text",
     "import_gds",
@@ -148,4 +144,4 @@ __all__ = (
 )
 
 
-__version__ = "7.1.3"
+__version__ = "7.2.0"