+++ gdsfactory/samples/07_flattening_device.py
@@ -21,9 +21,9 @@ from gdsfactory.component import Component
 def flatten_device() -> Component:
     c = gf.Component("test_remap_layers")
 
-    c.add_ref(gf.components.rectangle(size=(10, 1), layer=gf.LAYER.WG))
-    c.add_ref(gf.components.rectangle(size=(10, 2), layer=gf.LAYER.SLAB90))
-    c.add_ref(gf.components.rectangle(size=(10, 3), layer=gf.LAYER.SLAB150))
+    c.add_ref(gf.components.rectangle(size=(10, 1), layer=(1, 0)))
+    c.add_ref(gf.components.rectangle(size=(10, 2), layer=(3, 0)))
+    c.add_ref(gf.components.rectangle(size=(10, 3), layer=(2, 0)))
 
     assert len(c.references) == 3
     c2 = c.flatten()